Moved ClassroomModalForm to ClassroomModal/components/ClassroomForm.js Moved ClassroomModalForm to its own file, to help me see the code more easily when moving to new modal approach
