+++ client/src/pages/Dashboard/views/UserSettings/components/ClassroomModal/utils/useClassroomModalLoader.js
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import api from "utils/api";
 
 export const useClassroomModalLoader = (roomId) => {
 	const [room, setRoom] = useState();
